move test_egain_on_buffer_size
