+++ tests/unit/io/utils.py
@@ -13,12 +13,17 @@
 # limitations under the License.
 
 from cassandra.connection import HEADER_DIRECTION_TO_CLIENT
-from cassandra.marshal import uint8_pack, uint32_pack
+from cassandra.marshal import int32_pack, uint8_pack, uint32_pack
 from cassandra.protocol import (write_stringmultimap, write_int, write_string,
                                 SupportedMessage, ReadyMessage)
 
+import six
 from six import binary_type, BytesIO
 from mock import Mock
+
+import errno
+import os
+from socket import error as socket_error
 
 try:
     import unittest2 as unittest
@@ -213,3 +218,35 @@ class ReactorTestMixin(object):
 
         self.assertTrue(c.connected_event.is_set())
         return c
+
+    def test_egain_on_buffer_size(self):
+        c = self.test_successful_connection()
+
+        header = six.b('\x00\x00\x00\x00') + int32_pack(20000)
+        responses = [
+            header + (six.b('a') * (4096 - len(header))),
+            six.b('a') * 4096,
+            socket_error(errno.EAGAIN),
+            six.b('a') * 100,
+            socket_error(errno.EAGAIN)]
+
+        def side_effect(*args):
+            response = responses.pop(0)
+            if isinstance(response, socket_error):
+                raise response
+            else:
+                return response
+
+        self.get_socket(c).recv.side_effect = side_effect
+        c.handle_read(*self.null_handle_function_args)
+        self.assertEqual(c._current_frame.end_pos, 20000 + len(header))
+        # the EAGAIN prevents it from reading the last 100 bytes
+        c._iobuf.seek(0, os.SEEK_END)
+        pos = c._iobuf.tell()
+        self.assertEqual(pos, 4096 + 4096)
+
+        # now tell it to read the last 100 bytes
+        c.handle_read(*self.null_handle_function_args)
+        c._iobuf.seek(0, os.SEEK_END)
+        pos = c._iobuf.tell()
+        self.assertEqual(pos, 4096 + 4096 + 100)
